moonlight: secure mapping removing + creaet an empty mapping file before doing the configuration

+++ projects/configgen/configgen/generators/moonlight/moonlightControllers.py
@@ -42,7 +42,9 @@ def writeControllersConfig(system, rom, controllers):
     for controller in controllers:
         playerConfig = generateControllerConfig(controller, controllers[controller])
         confFile = recalboxFiles.moonlightMapping[int(controller)]
-        os.remove(confFile)
+        if os.path.isfile(confFile):
+            os.remove(confFile)
+        os.mknod(confFile)
         mappingFile = unixSettings.UnixSettings(confFile, ' ')
         mappingFile.save("# Device name", controllers[controller].realName)
         mappingFile.save("# SDL2 GUID  ", controllers[controller].guid)
